Fix header size validation and body size type mismatch for large uploads
  - Fix incorrect header size validation in readHeaders()
    The server was checking total buffer size (headers + body) against
    MAX_HEADER_SIZE limit, causing valid large body uploads to be
    rejected with "Headers too large" errors. Now validates only the
    header portion by checking for \r\n\r\n delimiter first, then
    measuring up to that point for complete headers or entire buffer
    for incomplete headers.

  - Change validateBodySize parameter type from int to unsigned long
    Ensures consistency with getContentLength() return type and properly
    supports large file uploads beyond 2GB without integer overflow.

  Fixes issue where 4MB+ uploads were incorrectly rejected despite being
  within client_max_body_size limits.

  This message clearly explains:
  1. What was broken (header validation rejecting valid uploads)
  2. Why it was broken (checking entire buffer instead of just headers)
  3. How it was fixed (check delimiter first, validate correct portion)
  4. Secondary fix (type consistency for large file support)
  5. The user-facing impact (4MB+ uploads now work)

+++ src/config/config/config.hpp
@@ -61,7 +61,7 @@ struct LocationConfig
 	// HTTP behavior
 	std::vector<std::string> allow_methods;
 	std::map<int, std::string> error_pages;
-	int client_max_body_size;
+	unsigned long client_max_body_size;
 
 	// CGI configuration
 	std::vector<std::string> cgi_path; // cgi_interpreters
@@ -106,7 +106,7 @@ struct ConfigData
 	// HTTP behavior
 	std::vector<std::string> allow_methods;
 	std::map<int, std::string> error_pages;
-	int client_max_body_size;
+	unsigned long client_max_body_size;
 
 	// CGI configuration
 	std::vector<std::string> cgi_path; // cgi_interpreters
